kb_mcp/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kb-mcp"
version = "0.2.0"
description = "Projection + MCP server exposing the ee-kb knowledge base to AI clients."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "ee-kb" }]
keywords = ["mcp", "knowledge-base", "embeddings", "search", "zotero"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Topic :: Text Processing :: Indexing",
]
dependencies = [
    "kb-core>=0.1.0",
    "mcp>=1.0.0",
    "python-frontmatter>=1.0.0",
    "PyYAML>=6.0",
    # Phase 2a: projection layer. sqlite-vec used in 2b, not 2a, but
    # declared here so a single `pip install -e .` covers both.
    "sqlite-vec>=0.1.3",
    # Phase 2b: embedding via OpenAI API. No torch / transformers —
    # everything cloud-side, keeping kb-mcp lightweight for server
    # deployment.
    "openai>=1.0.0",
]

[project.optional-dependencies]
# Phase 3b: write tools. kb_write is optional — without it, kb-mcp
# starts in read-only mode (write MCP tools unavailable).
write = [
    "kb-write>=0.1.0",
]
# Gemini embedding provider. Install with `pip install kb-mcp[gemini]`
# if you want to use GEMINI_API_KEY instead of OPENAI_API_KEY.
gemini = [
    "google-genai>=0.3.0",
]

[project.scripts]
kb-mcp = "kb_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/kb_mcp"]

[tool.hatch.build.targets.wheel.force-include]
# schema.sql ships inside the package so ensure_schema() can load it
# via importlib.resources regardless of install mode.
"src/kb_mcp/schema.sql" = "kb_mcp/schema.sql"

[tool.hatch.build.targets.sdist]
include = ["src/kb_mcp", "README.md", "pyproject.toml"]
